Split encounter by zone
This is an early alpha for splitting encounter by zone. Right now, it
takes the name of the zone when someone in the party enter the instance,
then go back to BNS zone when the dynamic quest is completed/the zone is
leaved. All encounters not in a dungeon will be classed under the bns
name.